+++ lib/declaration.ml
@@ -1918,6 +1918,10 @@ let read_interaction_value : type a.
   | Ms_user_select -> Some (v Ms_user_select (read_user_select t))
   | Webkit_text_fill_color -> Some (v Webkit_text_fill_color (read_color t))
   | Webkit_text_stroke_color -> Some (v Webkit_text_stroke_color (read_color t))
+  | Webkit_text_stroke_width ->
+      Some (v Webkit_text_stroke_width (read_border_width t))
+  | Webkit_text_stroke ->
+      Some (v Webkit_text_stroke (read_webkit_text_stroke t))
   | Moz_user_select -> Some (v Moz_user_select (read_user_select t))
   | Pointer_events -> Some (v Pointer_events (read_pointer_events t))
   | Resize -> Some (v Resize (read_resize t))